Rework openTreasure logic
The original openTreasure function was pretty interesting! I know almost
nothing about javascript/web-dev, but I made it more deterministic,
added a tiny bit of error checking, and move some duplicated code out of
the individual if substatements.

The if conditions are a little long and might not be super readable, so
just a suggestion!